Use dedicated go files for new structs
