+++ pkg/globalconfig/messages.go
@@ -1,0 +1,6 @@
+package globalconfig
+
+// MessagesConfig is the struct defining the messages config.
+type MessagesConfig struct {
+	TickerInterval int `yaml:"ticker_interval,omitempty"`
+}
